requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 8.2.0
matplotlib == 3.4.2
numpy == 1.20.3
scikit_image == 0.18.1
